Move parsing into submodule of registry
